# src/docs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Auth Service API Documentation",
    "description": "Documentation for the Authentication Service API"
  },
  "servers": [
    {
      "url": "http://localhost:3001/auth",
      "description": "Auth Endpoints"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Gateway Endpoints"
    }
  ],
...
  "paths": {
...
    "/vehicle/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/driver/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}